fix comments

+++ controllers/authControllers.js
@@ -4,26 +4,15 @@ import bcrypt from "bcrypt";
 import jwt from "jsonwebtoken";
 import dotenv from "dotenv";
 dotenv.config();
-import {
-  findUser,
-  emailUnique,
-  createUser,
-  setAvatar,
-} from "../services/authService.js";
+import { emailUnique, createUser, setAvatar } from "../services/authService.js";
 import fs from "fs/promises";
 import path from "path";
-import { resizeImage } from "../middlewares/imageHelpers.js";
-import { v4 as uuidv4 } from "uuid";
 import Jimp from "jimp";
-
-// const gravatar = require("gravatar");
 
 const { SECRET_KEY } = process.env;
 
-const tmpDir = path.resolve("tmp");
 const avatarsDir = path.resolve("public/avatars");
 
-// import * as authServices from "../services/authServices.js";
 export const register = async (req, res, next) => {
   const { name, email, password } = req.body;
   try {
@@ -31,7 +20,6 @@ export const register = async (req, res, next) => {
     if (user) {
       throw HttpError(409, "Email is already in use");
     }
-
     const newUser = await createUser({ ...req.body });
     const { avatarURL } = newUser;
     res.status(201).json({
@@ -96,12 +84,17 @@ export const logout = async (req, res, next) => {
 
 export const updateAvatar = async (req, res, next) => {
   try {
+    if (!req.file) {
+      return res
+        .status(400)
+        .json({ message: "File not found, please attach one" });
+    }
     const { _id } = req.user;
     const { path: oldPath, filename } = req.file;
 
     Jimp.read(oldPath, (err, lenna) => {
       if (err) throw err;
-      lenna.resize(250, 250).write(`${avatarsDir}\\${filename}`);
+      lenna.resize(250, 250).write(`${avatarsDir}/${filename}`);
       fs.rm(oldPath);
     });
     const avatarURL = path.join("avatars", filename);
